backend/requirements.txt
# Flask and extensions
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
Flask-Limiter==3.11.0

# Database
PyMySQL==1.1.1
cryptography==44.0.0

# Security
bcrypt==4.2.1

# Utils
python-dotenv==1.0.1
Flask-Caching==2.3.0
redis==5.2.1
